Rename component to NetworkBuilder.vue
