+++ components.d.ts
@@ -14,7 +14,7 @@ declare module 'vue' {
     DeleteTableDialog: typeof import('./src/components/DeleteTableDialog.vue')['default']
     DeleteWorkspaceDialog: typeof import('./src/components/DeleteWorkspaceDialog.vue')['default']
     DownloadDialog: typeof import('./src/components/DownloadDialog.vue')['default']
-    NetworkCreationTool: typeof import('./src/components/NetworkCreationTool.vue')['default']
+    NetworkBuilder: typeof import('./src/components/NetworkBuilder.vue')['default']
     NetworkPanel: typeof import('./src/components/NetworkPanel.vue')['default']
     PermissionsDialog: typeof import('./src/components/PermissionsDialog.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
@@ -61,7 +61,6 @@ declare module 'vue' {
     VOverlay: typeof import('vuetify/lib')['VOverlay']
     VProgressCircular: typeof import('vuetify/lib')['VProgressCircular']
     VProgressLinear: typeof import('vuetify/lib')['VProgressLinear']
-    VResponsive: typeof import('vuetify/lib')['VResponsive']
     VRow: typeof import('vuetify/lib')['VRow']
     VScrollXTransition: typeof import('vuetify/lib')['VScrollXTransition']
     VSelect: typeof import('vuetify/lib')['VSelect']
